Make find_recipe() return specific recipe.
find_recipe() was just returning the first recipe from the set, not the one we
were looking for. It now checks the recipe_uri and tries to find the recipe
with specified id reverting to previous behaviour if failed to do so.
Parsing recipe_uri is not the best way to do this so in future we should
probably track the recipe_id somewhere inside the server.