Update restraint client to use new streaming support
The connection from the client will now stay open during the entire execution
of the recipe and all results will be streamed back on that connection.

Tests that cause a reboot or panic are handled with a re-connect loop
which will keep trying to connect to the remote machine.  Once the machine
is back up it will continue the recipe where it left off.