Conflict with rhts-test-env instead of Providing it
If restraint-rhts Provides rhts-test-env then we can't include it in the
Beaker harness repos, because yum will pull it in instead of the real
rhts-test-env on jobs which are expecting to still use beah.

If people want to install restraint on their workstation, they will just
need to explicitly remove rhts-test-env first.

Change-Id: I05bd83442be5f109e999d860392d219af65498b2