use glib format specifier macros instead of C99 ones
This fixes a number of format-specifier warnings which break the build
in some circumstances (Fedora x86_64 with -m32, and RHEL4 S/390 31-bit
for reasons that are not totally clear to me).

I suspect this is actually a flaw in the glibc headers, but using the
glib macros is a better fit (since we are using the glib types
everywhere) and it makes portability glib's problem instead.

Change-Id: I0f8b62985688977ba17b1ee355259b475cbb11bf