third-party/openssl: skip building docs
I noticed this because the openssl build was failing on rawhide due to
missing /usr/bin/pod2man. We can just skip all the docs stuff entirely,
they are not needed here.

Unfortunately there's no configure switch to control it, we have to just
invoke the install_sw make target, which installs "software" (I guess)
but not docs.

Change-Id: If2a40b4f8a426d25e5b51cf9435cf84046c5ae90