Fix infinite uploadloop bug.
It is possible to make rstrnt-report-result (and most likely other chunks of
code that use upload_file function from src/upload.c) loop infinitely. To do so
all you need is redirect rstrnt-report-result's output to a file it is
uploading. Because upload_file prints a dot for each chunk it uploads the file
is never fully uploaded.

This is easily reproducible with any test that contains a line like this:
report_result $TEST FAIL 0 >> $OUTPUTFILE

Fix this by limiting upload to file's size at the start of upload.