add MIME type to links to log files
Add a type="text/plain" attribute to links to log files in the client
generated index.html.

Normally the web server sends a MIME type with a file, but when viewing
a local file:///index.html generated by restraint-client, there is no
web server to set the MIME type.  Thus, the index.html needs to set the
MIME type in the link, i.e.,
  <a href="resultoutputfile.log" type="text/plain">