Add relaxng validation.
Use libxml2's relaxng module to validate xml on load in client. A job schema
file is now included in client rpm. New restraint option --novalidate (-n) is
added to continue supporting xml's exported from beaker, with this option
any xml data that is not in the schema will be ignored.

Change-Id: Id611dbc13355da409d34a6b9df37b64c0a8efc1d