third-party: error out if tarballs fail to fetch
The -f option for curl makes it exit non-zero if the download failed.

The xz tarball is randomly failing to download sometimes, this won't
help but at least the build will fail sooner rather than going to Koji
and complaining about data not in gzip format.

Also added -L to all downloads to ensure we follow redirects.

Change-Id: I128740c10d6e52ea874d4ad9b512611c767e8d25