add softDependencies metadata
softDependencies are optional dependencies for a task.  If the
dependencies cannot be installed, the task will proceed anyway instead
of aborting as is done for (hard-)dependencies.

Example: LTP (Linux Test Project) can build NUMA tests if the numactl
and numactl-devel packages are installed, but the numactl packages are
not available on non-NUMA systems (like s390x), so the task should not
abort on non-NUMA systems.

Resolves feature-request RHBZ#1334476
  https://bugzilla.redhat.com/show_bug.cgi?id=1334476

Signed-off-by: Jeff Bastian <jbastian@redhat.com>

Change-Id: Ie84a1fe577d740e8100bf9dfb99303234b935701