Fix suicide by ssh.
If ssh-tunnel thread hasn't exited by the time we are cleaning up in the main
thread, main thread would send a SIGTERM by means of pthread_kill. The
assumption that this signal only affects the thread it is sent to doesn't hold
for some reason, resulting in mainthread being occasionally killed by sigterm.
Use pthread_cancel instead of pthread_kill to fix this.

Bug: 1353876