sync: adjust set timeouts.
On slower hosts, such as s390x 1 second is sometimes not enough for daemon to
come up and create a socket. Besides, if set fails, the job will fail anyway as
the other part would block indefinitely. So this commit changes set to wait for
the socket forever and changes delays from 500us to 50ms.

Change-Id: I249d66452830c12665c72018ac5fda3e19642084