Fix a segfault in openssl on ppc64.
When trying to fetch through https on ppc64 restraintd would crash with
an unintelligible backtrace pointing to SHA1_Update invocation. At the
same time when building on ppc64 openssl throws the following warning:

  Operating system: ppc64-whatever-linux2
  WARNING! If you wish to build 64-bit library, then you have to
           invoke './Configure linux-ppc64' *manually*.
           You have about 5 seconds to press Ctrl-C to abort.

This patch does exactly that, i.e. runs Configure with linux-ppc64
specified in case we are building on ppc64 machine.

Change-Id: Idf7b02bd24bfda1ecf807b4a91ce9b6364fdc4c8