Fix build faults in restraint RHEL4 workarounds.
The RHEL4 patch was getting applied to all builds due to the spec
file being wrong.

I've fixed the spec file so it will properly delete the RHEL4
patch on non-RHEL4 systems.

I've rejigged the patches so the RHEL4 workround for setpipe is
now in the glib-rhel4 patch.  This means that the RHEL4 patch is
dependent on the setpipe patch being applied first.

Change-Id: I235bb8db8090297aa10c108afad642d95fcaed15