Fix up printf arguments for report-result, make SCORE optional
The SCORE argument in rhts-report-result is optional, and restraint
has code to optionally upload it if it's set, but the arg count
tests actually make the score mandatory.  Fix so that score is
an optional argument.

The previous patch changed argument parsing to use glib to process
positional arguments, instead of using argv directly.

I missed one place where the argv arguments were still being used,
fix so that argv and argc are no longer used.

Bug: 1454325
Change-Id: I3e0cf3fc546a9746a4428adb8db0135959875b5e