Segfault in restraint_parse_url for host only git repositories
The hand rolled restraint_url code was buggy in that the
regex used to "parse" the URL was not extracting port
numbers at all.

Instead of fixing buggy code we can just remove it and
replace it with code that is known to work, namely LibSoup.

We already use LibSoup extensively so we're not adding
any new dependencies here.

Bug: 1542801
Change-Id: I12d88ec8a9e57a20aa8110f5a85f7f5ff4e09f4c