make rstrnt-report-result accept rhts-report-result style arguments
beakerlib and other code expects to be able to use rhts style args
for reporting results.

This patch will give rstrnt-report-result the ability to parse
arguments the same way as rhts-report-result does, preserving
any odd behavour that is expected from rhts-report-result.

rstrnt-report-result will detect the filename it was invoked as
and switch into rhts compatability mode if needed.

To do this the argument parsing code was split up into a common
block of code to parse environment variables, and then rhts and
rstrnt specific argument parsing codepaths.

The code that actually does the work has been split out into
an upload function that will take a processed argument struct
and operate off that. This isolates the upload code from the
argument parsing insanity.

The rhts-report-result compatability wrapper is no longer
needed and has been removed.  The new rhts-report-result
command is now part of the main restraint package.

Bug: 1571784
Change-Id: Idb815440128244c8d9b9a72e86c932567304900f