update Github URLs
Actually this is not the best place to point at either, since Github is
just a mirror and we typically do not publish source tarballs there
(they are produced by tito and then just... never put anywhere).

So this will need further tidying up in future.

But at least https://github.com/beaker-project/restraint is an
up-to-date copy of the source tree, and we will see any pull requests
that get posted there.

Change-Id: I4092f6e76b45d4e0342ff680aa3157fc6ac12000