Reset SIGPIPE action to default in child process.
glib2 will set SIGPIPE to be ignored when a GSocket object is
created.  When we fork() and exec() a child process we need to
reset SIGPIPE back to its default action so the exec'd processes
behave normally when sent a SIGPIPE.

To prevent any future problems of this nature, we will forcefully
set every signal back to its default value.

forkpty() will also need the same treatment.  As forkpty() is an
all-in-one fork() & exec() function in glibc, I have copied the code
from glibc and modified it to take a setup function to reset the
signal handlers.

Bug: 1592376
Change-Id: I6620a50abd373c2d1a16d04042f30a2ae9dd0b6d