Fix handing of gchar and xmlChar pointers.
The client code seems to freely mix up xmlChar and gchar pointers.

In addition to making the code a bit confusing as we can't be sure
of what type we're actually pointing to, it results in memory that
was allocated by libxml2 being passed to g_free instead of xmlFree.

I've cleaned up all the instances that I have found where gchar and
xmlChar pointers were mixed up.  There may be others lurking that
I have missed.

Change-Id: If16ae4c71a72a83d3b8b285e4f6fd47b318c2f1e