base package requirement should be fully versioned
We don't generally try to maintain any kind of backwards or forwards
compatibility between the restraint-rhts wrappers scripts and the
corresponding restraint code. So we really need to ensure that the
restraint and restraint-rhts packages always match exactly.

This will prevent yum from doing stuff like this, which may not work
properly (I asked it to install a specific version of restraint-rhts):

    ================================================================================
     Package         Arch    Version                          Repository       Size
    ================================================================================
    Installing:
     restraint-rhts  x86_64  0.1.35-1.el6bkr                  beaker-harness   49 k
    Installing for dependencies:
     restraint       x86_64  0.1.35-1.git.8.6df58ba.el6bkr    beaker-harness  7.6 M

    Transaction Summary
    ================================================================================
    Install       2 Package(s)

Change-Id: I3a22f0ca8e7dfa41afb99c6cd68305015e66968e