fix initscript startup on RHEL4
On RHEL4, the "daemon" shell function does not pass the command string
through a shell, it word-splits it and executes it directly. So the
subshell and redirection operators in the command do not work.

We can explicitly pass the command through bash -c instead.

This works fine on newer RHELs as well (where the command *is* passed to
a shell), the outer shell just exits and the inner one continues
running, so the net result is the same there.

Bug: 1585904
Change-Id: I7421c14c1dc143a6aad88e64501502b1167ba614