fix requirement on /usr/bin/hostname
It's still in /bin/hostname on RHEL7. It moved to /usr/bin/hostname as
part of the Fedora UsrMove and will presumably be in /usr/bin/hostname
on RHEL8+ as well.

Ultimately both requirements work, because of the special UsrMove
handling in the package manager -- but we want to correct requirement to
keep our repoclosure checks happy.

Change-Id: I991e72c584cdf602745916862570a69847e90581