tests: don't use Xvfb
It's not clear to me exactly what failure this test case is trying to
reproduce. But it seems like it should not be anything specific to Xvfb.
We can just put `sleep` in the background instead.

This way, we won't leave an Xvfb process hanging around, and we don't
need to skip the test if Xvfb is not installed.

Moreover this test was failing on my machine because Xvfb produces some
extra spew about keyboard configurations that breaks the expected output
in the test.

Change-Id: I3b58cad45f1f65376610804e3136da3a4859976e