RFC Restraint missing link to current tasks log file named 'current.log'
It was discovered that a file named current.log did not exist with restraint.
'current.log' is a link to each unique task file named /tmp/tmp.XXXX (where
XXXX is a random number) as the job runs.  At one point, this link was
used in the code for the beah harness but is no longer needed in
restraint code.  However, I think it's a user friendly feature to have
in restraint so adding the link back in. As the tasks change, the link
is updated to point the the tmp.XXX file for current task.

bug: 1609330
Change-Id: I2155b8946b42b6ae950a39799c11ec54a660557e