Remove libssh from restraint client
This changeset removes libssh from the restraint client.  The
restraint client sends configuration to server by way of stdin.
When run by beaker, the name of url containing the configuration
is defined in the config file ``/var/lib/restraint/config.conf``.

This changeset also allows the client to send a random config
filename for the server to use to checkpoint data.  This is to
prevent conflict with beaker initiated run as well as previous
client runs.  This file name is prefixed with 'checkpoint' and
suffixed with random string.  It is passed in the recipe property
data as 'checkpoint_file'. As server reboots, this file name is
retrieved from recipe configuration and used to recover checkpoint
data.

While testing this changeset, it was learned that restraint
is no longer produces chatty output in /var/log/messages.

Bug: 1667510
Bug: 1699795
Bug: 1653272
Bug: 1625990
Change-Id: Ica3ae4500297e1f2d564a9b903c85d95bee7d870