When local watchdog expires, upload task's log
When the local watchdog expires, it was not uploading the
current task's log file to beaker even though
rstrnt-report-result was being called.  This log file
is crucial to determine where the task is hung.
During debugging, I found OUTPUTFILE was not set.  Change
is to set OUTPUTFILE like in other places in code.  This
bug combined with 1609330 makes restraint more
backward-compatible with beah.

bug: 1700886
Change-Id: Ie96c9819b6568f20f5d1171da57a5b9ff4e763d6