Restraint to support fail/false strings files
To remain backward compatible to beah, it was suggested
to support use of beah named files 'failurestrings' and
'falsestring' in the directory /usr/share/rhts.
If these files are created by the user, they
will be used over the hard coded defaults.  The
environment variables will take precedence
over these files and the hardcoded defaults.

The benefit to supporting these files is they only need
to be defined once as opposed to injecting the same
values as environment variables for each task.
Additionally, the files can be unique for each project
as opposed to hardcoded defaults.

Bug: 1613164
Change-Id: I2786d89c8f2a21ea6a3d04af68bd54683bd05402