Random results with rstrnt-adjust-watchdog cmd
User found latest code with BZ1700926 was failing when
using 600 seconds but successful with 1200 seconds.

An additional change with BZ1700926 was to add 30 minutes to the EWD
when rstrnt-adjust-watchdog is executed.  This is so it behaves similar
to what we do when a task is started. This change is what caused
things to break.

In triaging, the restraintd server sent message with 2400 seconds
(600 as requested from rstrnt-adjust-watchdog plus EWD adjustment of
30 minutes/1800 seconds).  While running beaker client command, I
saw 'bkr show-watchdog <task=id> return 240 instead of 2400. So
it looks like 1 byte is truncated from the message.

The command rstrnt-adjust-watchdog sends the message with 600 seconds.
When restraintd server receives this message, it adds 30 minutes to the same
message received and that message is cloned and sent to lab controller.
The problem is that the message header len is not adjusted and
message body now contains "2400" instead of "600" (1 more byte).
Initializing message header length to 0 forces the soup library to
reset the header length before it sends the message to lab controller.

Bug: 1753336
Change-Id: I93911aba2e609f3af612a095bfa8a4ccd98e1e8d