Plugin 01_dmesg_check fails in rlse 0.1.40
Simple debug code was written which showed the file
/usr/share/rhts/failurestrings existed but it was empty.
The solution to this issue is to make sure the file
contains data to consider whether to use it.

The extra code written in this changeset is debug code
to give us visibility to triage issues more efficiently.

Debug code is written at the bottom of resultoutputfile.log when
01_dmesg_check reports failure.  This debug code reports which
set of failure and false string data was used: environment vars,
files, or hardcoded defaults.  It shows content of the
failure/falsestrings variables and if the files exist, if there
is data in them or the files content is also dumped into the
bottom of the log file.

Bug: 1751074
Change-Id: I64a6be9df02ece738d24592b9da874c2a106236a