Rstrnt Client not provide task vers in job.xml
It was reported that restraint client does not provide
the task test 'version' in the job.xml.  This is also
true for Beah so is not restraint specific.  In close
inspection of beaker recipe xml output, only tasks with
RPMs contain version numbers.  Those with fetch operations
do not.

This changeset allows restraint client to write tasks
in job.xml output which include version number from
RPM tasks only.  This is acquired from the rpm itself.
The server returns the version number (if available) in
the tasks 'Completed|Aborted' status message to the client.
The restraint client writes it out in the job.xml

At this time, this changeset does not affect Beaker.
No changes at this time are made for fetch url operations.

Bug: 1691485
Change-Id: Ic36c70b32bd3b6f4c70f8409ccfb798a5ad23495