Seeing too many repo extractions
Client reported seeing many repo extractions when fetching
their repo.  They were using repoRequires in their Makefile.
Once this directory was extracted, it resulted in the condition
"WARNING: No metadata for dependency" since the directory
didn't even have a Makefile or testinfo.desc or metadata file.
The code then proceeded to do another fetch operation on the
same directory when it should have gone to the next
repoRequires in the list as done in other places in the code.
The change is to get next repodeps in list.

Bug: 1774211
Change-Id: If80c93e4385be8a29e005588bac0bc7cc0310d9c