Task Fetch https generates noisy log file
When using fetch https operation, it generates many log entries with
** Extracting filename
It does for every file/directory found in the repo when it should
only report on the directory requested.  In other words, if the
following is used in the task:
<fetch url="https://gitlab.com/user/reponame/~/archive/master/reponame-master.tar.gz#reponame-master/mydir"/>
Then  we should only log what is found beneath mydir and not all files
beneath reponame.  This is causing performance issues when the repo
is large.

Bug: 1782422
Change-Id: Ia3d0dd64c6d1ed3d9d10bad1755297d6f7fb2de1