Implement cache for harness and task logs
This commit implements a caching mechanism for logs to avoid network
round trips for every time the child process writes to stdout. First,
all logs are written to local files in a separate thread, which only
requires to block the task execution at the end, when we wait for the
remaining operations to continue and queue the HTTP request. However,
with this method, error handling becomes more complicated, in that
everything is fire-and-forget.