Resolve Memory leak in upload_file()
Caller of g_uri_escape_string is responsible for freeing the returned string,
but the function was called as actual argument of soup_uri_new_with_base,
loosing the reference to the string.  Keep the reference to the string
returned by g_uri_escape_string and free it after soup_uri_new_with_base
is called.

The valgrind definitely leak error leak seen is as follows:
==32571== 16 bytes in 1 blocks are definitely lost in loss record 687 of 2,056
==32571==    at 0x4C29EBD: malloc (vg_replace_malloc.c:308)
==32571==    by 0x4C2C210: realloc (vg_replace_malloc.c:836)
==32571==    by 0x54C78D: g_realloc (gmem.c:164)
==32571==    by 0x562B63: g_string_maybe_expand (gstring.c:102)
==32571==    by 0x562BB1: g_string_sized_new (gstring.c:127)
==32571==    by 0x5733F5: g_uri_escape_string (gurifuncs.c:247)
==32571==    by 0x40754B: upload_file (upload.c:108)
==32571==    by 0x406C39: main (cmd_log_main.c:56)

Bug: 1813055
Change-Id: Ibed678cd4d224b51b98df27cc7b0f4211bc46a8b