Revert to stderr to bring console output back
Part of the changes made for 'BZ1667510 - remove sshlib' eliminated
restraint's console output which many users rely on.  This changeset
aims to provide what users come to expect by reverting the changes
that removed the console output.  We will later reconvene
with users to determine what minimum output they need so we can
minimize what is output onto the console.

Bug: 1815570
Change-Id: Iaf1ef24532ed8315efa65e4e6892b4e7fb313b09