tests: Cover no rpm program in get_package_version
Add test to make sure that get_package_version handles the error when
the rpm program is not available.

Update existing tests where rpm is required to be skipped if the rpm
program is not available. This ensures that the suite can be run on
systems where rpm is not available.