Fix: RSTRNT to create well-known OUTPUTFILE var
In order to eliminate dependency of rhts_environment.h, restraint
should create OUTPUTFILE environment variable used by the command
rstrnt-report-result.  This is done once for each task execution
when the user's task is run.  If the local watchdog is executed,
it will see this file and if there is data will report the results.

Bug: 1813946