BLOCKED: Report Task Abort Status when Local Watchdog Expires
BLOCKED: Pending performance improvement fixes
When the local watchdog expires, it should report a task status of
Aborted.  Instead it shows Warn which is the last report result state.
Additionally, you should continue to see 10_localwatchdog and
99_reboot plugins in the "Task Result" section of the job.  Once
the machine reboots (due to lwd expire), it should continue onto
the next task.

The purpose of this change is to allow users to identify a failed
test.  Investigating this issue revealed documentation which
says Aborted is the desired result for LWD.  The code also was
leaning towards reporting a status of Aborted. However,
the 99_reboot doesn't allow execution to go back to restraintd
so reporting status of aborted does not complete.

Bug: 1804409