Fix use of uninitialized fd
The process_data->fd_in was left uninitialized when the process was
forked with pty. As process_data is allocated with g_slice_new0, the
default value for fd_in is 0, and it was used without being set during
fork.

Initialize fd_in and fd_out in process_data to -1.

Only use process->fd_in if it's set during fork.