Fix incorrect use of buffer
The buffer from g_io_channel_read was used directlty in g_debug
function. This is not correct because g_debug expects a NULL terminated
string, but there is no guarantee that the data returned by
g_io_channel_read is going to meet this requirement.
This will print everything until the first NULL character in the buffer,
regardless of how much data was read.

Replace g_debug call with fwrite in G_IO_STATUS_NORMAL case. The code is
borrowed from 0.1.45.