Fix: Prevent loop in local watchdog plugin
When the local watchdog (LWD) expires a task, the LWD plugin `20_sysinfo`
goes into an infinite loop since the directory `/mnt/testarea` is not
created for the non-rhts restraint package. An error returned by `diff`
utility within an infinite loop was not anticipated.  The fix
terminates the infinite loop when diff returns error.

Bug: 1813935