fix: free SoupURIs from soup_server_get_uris call
The soup_server_get_uris returns a list of SoupURIS. The list itself and
the SoupURIs in it should be freed by the caller. There is a hint for
this in GLibs docs, where the function return states "transfer full".

Use g_list_free_full to free SoupURIs and the list itself.