fix: json_to_hashtable frees json_object's hashtable key
The hash table returned by json_to_hastable function was created using
g_free to destroy the memory allocated for the key. The problem was that
the key used for each entry was the (char *) key from the json_object's
allocated memory.
Once the hash table is destroyed, the json_object will get corrupted,
and freeing the object will cause a double free.

Don't use key_destroy_func for the hash table returned by
json_to_hashtable function. The key will be free when the json_object is
freed.