Fix: Stop logging `LWD is disabled` every minute
When LWD (Local Watchdog) is disabled, there is a message in the
harness log that reports this every minute.  The message looks
like: `Localwatchdog at:  Disabled! `.  This changeset makes sure
it is no longer reported repeatedly when `no_localwatchdog=true`
is configured in the task `metadata` file.
To make sure the connection between client and server don't timeout,
the ssh arguments `ServerAliveInterval` and `ServerAliveCountMax` are
set on the `restraint` client. The `ServerAliveInterval` is set to
60 seconds while `ServerAliveCountMax` is set fo 5.  This results in
an idle timeout of 5 minutes.  The `ServerAliveCountMax` can be
altered by the user using restraint client `--timeout` argument.
This argument is valid only when rsh option is not used.
Due to LGTM checker errors, function localtime() was replaced
with localtime_r().

Bug: 1598329