fix: Noisy output from callback to client
Output for the client is sent by restraintd through the
restraint_stdout_message callback in connections_write. Printing the
messages to stdout was duplicating the messages in the client and
breaking the verbosity levels.

With the buffers filled to 0, reading to buffer size - 1 will ensure
that they are always NULL terminated, and can be use safely with
functions that expect a NULL terminated string.

In the task callback, suppress all output by using g_debug, that only
outputs when G_MESSAGES_DEBUG is used.

In the server callback, only print to stdout when restraintd is not
running through the client.