restraint systemd: Have pre-check script wait for IP address
Add a check to make sure the system has an IP address to the pre-check
script (/usr/bin/check_beaker) that is run by systemd.

This check will try for a maximum of 60 seconds to see if the system
has acquired an IP address. If the system has an IP address it will
continue. If after 60 seconds it still doesn't have an IP address it
will continue.

Also increase start timeout to 3 minutes (180 seconds). The default
appears to be 90 seconds.

There was an issue found where Debian 10.3 boots up but takes a longer
than expected time to get an IP address from DHCP. This caused
restraint to time out and fail to fetch the recipe file.

Fixes: #101