fix: time_chged leak in server_recipe_callback
A request to adjust watchdog will take up to HEARTBEAT time to get the
value updated. As time_chged is freed and set to NULL only after the
value is updated, every request to adjust the watchdog sent before this
will cause a leak.

Only allocate new memory for time_chged if the variable is NULL.