FIX: `SKIP` results fail for non-rhts tasks
When the task reports just `SKIP` for results, the final task result
should be `SKIP`. When a non-rhts task is running, an extra task result
of `PASS` is occurring. Since `PASS` has a higher severity than `SKIP`,
`PASS` becomes the final task result. (note: a non-rhts task is one
defined using `metadata` file instead of `Makefile/testinfo` file.)
Bugzilla 1334893 made a change to always report results `PASS` or
`FAIL` for non-rhts tasks based on the process exit results without
considering results from user calls to `rstrnt-report-result`.

New Behavior when task process exits successfully
-------------------------------------------------
Code changes monitor whether the user reports results by way of
`rstrnt-report-result`.  If so, give user `reported` results
priority; otherwise, provide a `PASS` task result for user.

Behavior when process does not exit successfully
------------------------------------------------
The intent of Bugzilla 1334893 was to make sure there is a final
task result. Normally when a non-zero process result is returned,
an error message is generated.  When the task is processed in
the COMPLETED state, this is translated as an abort.  For non-rhts
the result instead is FAIL.  This gives the user the option
to continue their job; otherwise, the user will
have to call rstrnt-abort is they want legacy behavior to stop
executing the remainder of the job.

Bug: 1855013