feat: Add ability to adjust log upload interval
Allow to adjust the upload interval for the log manager through the
configuration file "/var/lib/restraint/log_manager.conf".
The file uses the ini format, section "log-manager" and
the integer key "upload_interval".

Setting "upload_interval=0" disables the log manager. Current valid
range is between 3 and 60, both included. Setting a value outside this
range will be overridden with the closest valid value in the range.

Note that this feature is provided for debugging purposes only.