Fix error handling in restraint config functions
Previously, the config functions didn't check errors on key file reading. In case of error, the key file structure was left either empty or partially loaded.

Other than the file not found error, which is tolerated on purpose, doesn't make much sense to ignore other errors, and the key file structure shouldn't be used at all after an error.

The case of partially loaded file could be the root cause in bugzilla: 1783283, although the state of the config file at that point is still unknown.
This is not going to fix a bad configuration file, but at least the error will get visibility.

Add error handling for reading key files
Make invalid type errors critical
Print warning in connections write in case of errors in config_set

Note:  This pull request replaces pull #178 which the person who authored left the company
and we lacked privileges to rebase #178 changeset for success of checks.

Bug: 1783283