rstrnt-prepare-reboot: Make the BootNext reset timer optional
This script is called from rstrnt-reboot which is used for ordinary reboots
only, so it does not make any sense to reset BootNext in this case.  Moreover,
if the shutdown takes more than 180s, a completely different entry might be
booted instead (e.g. PXE instead of RHEL) making the executed job to timeout
and subsequently fail.