Extend task_fetch retries
More beaker jobs are now relying on fetching tasks from git repos, and intermitted connectivity issue are causing unnecessary problems when restraint only tries a couple of times in half a minute.
It's a much bigger issue when core tasks like kernelinstall was aborted for that reason. The remaining tests in the recipe become pointless and even misleading due to false positive/negative from wrong kernel.
This minor change is increasing the number and interval of task_fetch retries - to make future jobs more tolerant with intermitted connectivity issue.