Fix bug that legacy test case does not work when using 'metadata'
Restraint works in non-rhts_compat mode after added metadata file
in legacy test case, that implies that all ENV variable names
changed. but legacy test code still use old ENV names.

With this patch to ensure that the original test cases can work
regardless of whether it is in compat mode or not.

Signed-off-by: Jianhong Yin <jiyin@redhat.com>