Add multi machine CI infrastructure
Add infrastructure to allow testing of the restraint client and daemon
across two containers networked together.

Most of the changes are to support the initial bring up of two
containers and allow them to communicate over ssh.  There is a single
job xml to verify the infrastructure is working correctly.

Initial work is for two local containers to run.  Followon work will
support running in GitHub CI.

Signed-off-by: Don Zickus <dzickus@redhat.com>