Add support for rhel-10 and centos-stream-10
This patch fixes multiple build issues with CentOS-Stream 10 and
RHEL-10.  The first is using pcre2-static instead of pcre-static.  The
second is to handle a ppc failure that seems related to the
build_type_safety_check_c put in for Fedora > 40.  RHEL-10 is inheriting
the same problem.

RHEL-10 is also using gcc15 which is complaining about
incompatible-pointer types for libsoup code.  This was fixed 7 years
ago, so backport that patch here.

Packit and centos-stream10 is skipped for now because packit does not
build statically and instead relies on the -devel packages.
libsoup-devel is not available for centos-stream10 as they have moved to
libsoup3-devel and restraint isn't ported to it yet.

Review-checks will get centos-stream10 once EPEL builds tito which is
needed to trigger this workflow.

Signed-off-by: Don Zickus <dzickus@redhat.com>