Add support for rhel-10 and centos-stream-10
This patch fixes two build issues with CentOS-Stream 10.  The first is
using pcre2-static instead of pcre-static.  The second is to handle a
ppc failure that seems related to the build_type_safety_check_c put in
for Fedora > 40.  RHEL-10 is inheriting the same problem.

Update packit.yml file to include CentOS-Stream-10.

Signed-off-by: Don Zickus <dzickus@redhat.com>