Handle Fedora's new certificate file
Fedora removed the ca-certificates.crt file which is libcurl's default
certificate file and replaced it with
/etc/pki/ca-trust/extracted/pem/tls-ca-bundle.pem for Fedora 41 and
later.

Tell curl to build with the new location but only for later Fedora and
RHEL releases.  Use a text file as a trigger mechanism that is removed
by the spec file for older releases.  Mimic's an exist glib patch hack.

See https://fedoraproject.org/wiki/Changes/droppingOfCertPemFile

Signed-off-by: Don Zickus <dzickus@redhat.com>